+++ static/app/views/starfish/views/webServiceView/starfishLanding.tsx
@@ -1,5 +1,5 @@
-import {Fragment, useEffect, useRef} from 'react';
-import {browserHistory, InjectedRouter} from 'react-router';
+import {Fragment} from 'react';
+import {InjectedRouter} from 'react-router';
 import styled from '@emotion/styled';
 import {Location} from 'history';
 
@@ -7,13 +7,11 @@ import DatePageFilter from 'sentry/components/datePageFilter';
 import * as Layout from 'sentry/components/layouts/thirds';
 import LoadingIndicator from 'sentry/components/loadingIndicator';
 import PageFilterBar from 'sentry/components/organizations/pageFilterBar';
-import {PageHeadingQuestionTooltip} from 'sentry/components/pageHeadingQuestionTooltip';
 import * as TeamKeyTransactionManager from 'sentry/components/performance/teamKeyTransactionsManager';
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 import {Organization, PageFilters, Project} from 'sentry/types';
 import EventView from 'sentry/utils/discover/eventView';
-import {GenericQueryBatcher} from 'sentry/utils/performance/contexts/genericQueryBatcher';
 import {
   PageErrorAlert,
   PageErrorProvider,
@@ -25,68 +23,34 @@ import {StarfishView} from './starfishView';
 type Props = {
   eventView: EventView;
   location: Location;
-  onboardingProject: Project | undefined;
   organization: Organization;
   projects: Project[];
   router: InjectedRouter;
   selection: PageFilters;
-  setError: (msg: string | undefined) => void;
   withStaticFilters: boolean;
 };
 
 export function StarfishLanding(props: Props) {
-  const {organization, location, eventView, onboardingProject} = props;
+  const {organization, eventView} = props;
 
   const {teams, initiallyLoaded} = useTeams({provideUserTeams: true});
 
-  const hasMounted = useRef(false);
-  const showOnboarding = onboardingProject !== undefined;
-
-  useEffect(() => {
-    if (hasMounted.current) {
-      browserHistory.replace({
-        pathname: location.pathname,
-        query: {
-          ...location.query,
-          landingDisplay: undefined,
-        },
-      });
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [eventView.project.join('.')]);
-
-  useEffect(() => {
-    hasMounted.current = true;
-  }, []);
-
-  let pageFilters: React.ReactNode = (
+  const pageFilters: React.ReactNode = (
     <PageFilterBar condensed>
       <DatePageFilter alignDropdown="left" />
     </PageFilterBar>
   );
-
-  if (showOnboarding) {
-    pageFilters = <SearchContainerWithFilter>{pageFilters}</SearchContainerWithFilter>;
-  }
 
   return (
     <Layout.Page>
       <PageErrorProvider>
         <Layout.Header>
           <Layout.HeaderContent>
-            <Layout.Title>
-              {t('Starfish')}
-              <PageHeadingQuestionTooltip
-                docsUrl="https://docs.sentry.io/product/performance/"
-                title={t(
-                  'Your main view for transaction data with graphs that visualize transactions or trends, as well as a table where you can drill down on individual transactions.'
-                )}
-              />
-            </Layout.Title>
+            <Layout.Title>{t('Starfish')}</Layout.Title>
           </Layout.HeaderContent>
         </Layout.Header>
 
-        <Layout.Body data-test-id="performance-landing-body">
+        <Layout.Body>
           <Layout.Main fullWidth>
             <PageErrorAlert />
             <Fragment>
@@ -101,9 +65,7 @@ export function StarfishLanding(props: Props) {
                   selectedTeams={['myteams']}
                   selectedProjects={eventView.project.map(String)}
                 >
-                  <GenericQueryBatcher>
-                    <StarfishView {...props} />
-                  </GenericQueryBatcher>
+                  <StarfishView {...props} />
                 </TeamKeyTransactionManager.Provider>
               ) : (
                 <LoadingIndicator />
@@ -115,18 +77,6 @@ export function StarfishLanding(props: Props) {
     </Layout.Page>
   );
 }
-
-const SearchContainerWithFilter = styled('div')`
-  display: grid;
-  grid-template-rows: auto auto;
-  gap: ${space(2)};
-  margin-bottom: ${space(2)};
-
-  @media (min-width: ${p => p.theme.breakpoints.small}) {
-    grid-template-rows: auto;
-    grid-template-columns: auto 1fr;
-  }
-`;
 
 const SearchContainerWithFilterAndMetrics = styled('div')`
   display: grid;
